restControllers

+++ src/main/java/ma/ensaj/hotel/service/ReservationService.java
@@ -5,22 +5,43 @@ import ma.ensaj.hotel.repository.ReservationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
 public class ReservationService {
+
     @Autowired
     private ReservationRepository reservationRepository;
 
-    public Reservation saveReservation(Reservation reservation) {
-        return reservationRepository.save(reservation);
+    public List<Reservation> getAllReservations() {
+        return reservationRepository.findAll();
     }
 
     public Optional<Reservation> getReservationById(Long id) {
         return reservationRepository.findById(id);
     }
 
-    public void deleteReservation(Long id) {
-        reservationRepository.deleteById(id);
+    public Reservation saveReservation(Reservation reservation) {
+        return reservationRepository.save(reservation);
+    }
+
+    public Optional<Reservation> updateReservation(Long id, Reservation reservation) {
+        return reservationRepository.findById(id).map(existingReservation -> {
+            existingReservation.setClient(reservation.getClient());
+            existingReservation.setChambre(reservation.getChambre());
+            existingReservation.setDateDebut(reservation.getDateDebut());
+            existingReservation.setDateFin(reservation.getDateFin());
+            existingReservation.setPreferences(reservation.getPreferences());
+            return reservationRepository.save(existingReservation);
+        });
+    }
+
+    public boolean deleteReservation(Long id) {
+        if (reservationRepository.existsById(id)) {
+            reservationRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
